Move clearable thread pool out of wrong folder, update string util
