+++ recaf-ui/build.gradle
@@ -1,6 +1,9 @@
 plugins {
     id 'com.github.johnrengelman.shadow' version '6.1.0'
+    id 'application'
 }
+
+mainClassName = "me.coley.recaf.RecafUI"
 
 configurations {
     common
